Server/requirements.txt
aiohttp==3.7.3
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
collection==0.1.6
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.6.0
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==5.8.0
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.6.2
notebook==6.1.6
numpy==1.20.3
oauthlib==3.1.0
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
# pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
telepot==12.7
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
